+++ docs/next.config.mjs
@@ -6,7 +6,6 @@ const withMDX = createMDX();
 /** @type {import('next').NextConfig} */
 const config = {
   reactStrictMode: true,
-  output: 'export',
   images: {
     unoptimized: true,
   },