+++ docs/src/app/api/search/route.ts
@@ -1,10 +1,9 @@
 import { source } from '@/lib/source';
 import { createFromSource } from 'fumadocs-core/search/server';
 
-// Statically cached for static export
 export const revalidate = false;
 
-export const { staticGET: GET } = createFromSource(source, {
+export const { GET } = createFromSource(source, {
   // https://docs.orama.com/docs/orama-js/supported-languages
   language: 'english',
 });