+++ docs/src/components/api-page.client.tsx
@@ -1,4 +1,8 @@
 'use client';
 import { defineClientConfig } from 'fumadocs-openapi/ui/client';
 
-export default defineClientConfig({});
+export default defineClientConfig({
+  playground: {
+    requestTimeout: 30,
+  },
+});